All files are now at root dir
